+++ app/javascript/components/Login.js
@@ -1,0 +1,10 @@
+import React from "react"
+import PropTypes from "prop-types"
+class Login extends React.Component {
+  render() {
+    return (
+    );
+  }
+}
+
+export default Login
