--- app/javascript/components/Signup.js
@@ -1,55 +1,0 @@
-import React from "react"
-import PropTypes from "prop-types"
-class Signup extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      isModalOpen: false,
-      user_name: '',
-      user_email: '',
-      user_password: '',
-      user_password_confirmation: '',
-    };
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-  handleChange(event) {
-    this.setState({
-      [event.target.id]: event.target.value,
-    });
-  }
-
-  handleSubmit(event) {
-    alert('A form is submitted: ');
-    event.preventDefault();
-  }
-
-  render() {
-    return (
-      <form onSubmit={this.handleSubmit}
-        action={this.props.url} acceptCharset="UTF-8" method="post">
-
-        <input type="hidden" name="authenticity_token"
-          value={this.props.authenticityToken} />
-
-        <label htmlFor="user_name">Name:</label>
-        <input className="form" type="text" name="user[name]" id="user_name" value={this.state.user_name} onChange={this.handleChange} />
-
-        <label htmlFor="user_email">Email:</label>
-        <input className="form" type="text" name="user[email]" id="user_email" value={this.state.user_email} onChange={this.handleChange} />
-        <label htmlFor="user_password">password</label>
-        <input className="form" type="password" name="user[password]" id="user_password" value={this.state.user_password} onChange={this.handleChange} />
-        <label htmlFor="user_password_confirmation">password_confirmation</label>
-        <input className="form" type="password" name="user[password_confirmation]" id="user_password_confirmation" value={this.state.user_password_confirmation} onChange={this.handleChange} />
-
-        <input className="form_submit" type="submit" name="submit" value="Submit" />
-      </form>
-    );
-  }
-}
-
-Signup.propTypes = {
-  url: PropTypes.string,
-  authenticity_Token: PropTypes.string
-};
-export default Signup
